refactor(mastering): extract Stages 5.5/6/6.5/6.7 into _album_stages (#290 D5)
