+++ servers/bitwize-music-server/handlers/processing/audio.py
@@ -709,287 +709,21 @@ async def master_album(
     ):
         return result
 
-    verify_results = ctx.verify_results
-
     # --- Stage 5.5: Mastering samples (codec preview + mono fold-down QC) ---
-    # Issue #296. Writes .aac.m4a and .MONO_FOLD.md sidecars to the
-    # mastering_samples/ sibling directory so mastered/ stays WAV-only. A
-    # mono-fold hard-fail short-circuits the pipeline; codec preview never blocks.
-    from tools.mastering.master_tracks import GENRE_PRESETS, _PRESET_DEFAULTS
-
-    if genre and genre.lower() in GENRE_PRESETS:
-        sample_cfg: dict[str, Any] = dict(GENRE_PRESETS[genre.lower()])
-    else:
-        sample_cfg = dict(_PRESET_DEFAULTS)
-
-    codec_enabled = bool(int(sample_cfg.get("codec_preview_enabled", 1)))
-    codec_bitrate = int(sample_cfg.get("codec_preview_bitrate_kbps", 128))
-    monofold_enabled = bool(int(sample_cfg.get("mono_fold_enabled", 1)))
-    monofold_write_audio = bool(int(sample_cfg.get("mono_fold_write_audio", 1)))
-    monofold_thresholds = {
-        "band_drop_fail_db": float(sample_cfg.get("mono_fold_band_drop_fail_db", 6.0)),
-        "lufs_warn_db": float(sample_cfg.get("mono_fold_lufs_warn_db", 3.0)),
-        "vocal_warn_db": float(sample_cfg.get("mono_fold_vocal_warn_db", 2.0)),
-        "correlation_warn": float(sample_cfg.get("mono_fold_correlation_warn", 0.3)),
-    }
-
-    samples_dir = audio_dir / "mastering_samples"
-    samples_stage: dict[str, Any] = {
-        "status": "pass",
-        "codec_preview_enabled": codec_enabled,
-        "mono_fold_enabled": monofold_enabled,
-        "output_dir": str(samples_dir),
-    }
-
-    if codec_enabled or monofold_enabled:
-        samples_dir.mkdir(exist_ok=True)
-
-    # Codec preview — never blocks
-    if codec_enabled:
-        from tools.mastering.codec_preview import (
-            CodecPreviewError,
-            render_aac_preview,
-        )
-
-        codec_results: list[dict[str, Any]] = []
-        codec_errors: list[str] = []
-        for wav in mastered_files:
-            out_path = samples_dir / f"{wav.stem}.aac.m4a"
-            try:
-                info = await loop.run_in_executor(
-                    None, render_aac_preview, wav, out_path, codec_bitrate
-                )
-                codec_results.append({
-                    "track": wav.name,
-                    "output_path": info["output_path"],
-                    "bitrate_kbps": info["bitrate_kbps"],
-                })
-            except CodecPreviewError as e:
-                codec_errors.append(f"{wav.name}: {e}")
-                warnings.append(f"Codec preview {wav.name}: {e}")
-
-        samples_stage["codec_previews"] = codec_results
-        if codec_errors:
-            samples_stage["codec_errors"] = codec_errors
-
-    # Mono fold-down QC — hard-fails the pipeline on band drop
-    if monofold_enabled:
-        import soundfile as sf
-        from tools.mastering.mono_fold import mono_fold_metrics
-        from tools.mastering.mono_fold_report import render_mono_fold_markdown
-
-        def _do_mono_fold(wav_path: Path) -> dict[str, Any]:
-            data, rate = sf.read(str(wav_path))
-            import numpy as _np
-            if data.ndim == 1:
-                data = _np.column_stack([data, data])
-            metrics = mono_fold_metrics(data, rate, thresholds=monofold_thresholds)
-
-            stem = wav_path.stem
-            sample_filename = f"{stem}.mono.wav" if monofold_write_audio else None
-            if sample_filename:
-                sf.write(str(samples_dir / sample_filename), metrics["mono_audio"], rate, subtype="PCM_24")
-
-            md = render_mono_fold_markdown(stem, metrics, sample_filename)
-            (samples_dir / f"{stem}.MONO_FOLD.md").write_text(md, encoding="utf-8")
-
-            return {
-                "track": wav_path.name,
-                "verdict": metrics["verdict"],
-                "band_drop_fail": metrics["band_drop_fail"],
-                "worst_band": metrics["worst_band"],
-                "lufs_delta_db": metrics["lufs"]["delta_db"],
-                "vocal_delta_db": metrics["vocal_rms"]["delta_db"],
-                "stereo_correlation": metrics["stereo_correlation"],
-                "report_path": str(samples_dir / f"{stem}.MONO_FOLD.md"),
-            }
-
-        mono_results = []
-        for wav in mastered_files:
-            mono_results.append(await loop.run_in_executor(None, _do_mono_fold, wav))
-
-        mono_passed = sum(1 for r in mono_results if r["verdict"] == "PASS")
-        mono_warned = sum(1 for r in mono_results if r["verdict"] == "WARN")
-        mono_failed = sum(1 for r in mono_results if r["verdict"] == "FAIL")
-
-        for r in mono_results:
-            if r["verdict"] == "WARN":
-                warnings.append(
-                    f"Mono fold {r['track']}: WARN — see {Path(r['report_path']).name}"
-                )
-
-        samples_stage["mono_fold"] = {
-            "tracks": mono_results,
-            "passed": mono_passed,
-            "warned": mono_warned,
-            "failed": mono_failed,
-        }
-
-        if mono_failed > 0:
-            failed_tracks = [r for r in mono_results if r["verdict"] == "FAIL"]
-            samples_stage["status"] = "fail"
-            stages["mastering_samples"] = samples_stage
-            return _safe_json({
-                "album_slug": album_slug,
-                "stage_reached": "mastering_samples",
-                "stages": stages,
-                "settings": settings,
-                "warnings": warnings,
-                "failed_stage": "mastering_samples",
-                "failure_detail": {
-                    "reason": "Mono fold-down hard-fail (phase cancellation)",
-                    "tracks_failed": [r["track"] for r in failed_tracks],
-                    "details": [
-                        {
-                            "track": r["track"],
-                            "worst_band": r["worst_band"],
-                            "report": r["report_path"],
-                        }
-                        for r in failed_tracks
-                    ],
-                },
-            })
-
-    stages["mastering_samples"] = samples_stage
+    if result := await _album_stages._stage_mastering_samples(ctx):
+        return result
 
     # --- Stage 6: Post-QC ---
-    from tools.mastering.qc_tracks import qc_track
-
-    post_qc_results = []
-    for wav in mastered_files:
-        result = await loop.run_in_executor(None, qc_track, str(wav), None)
-        post_qc_results.append(result)
-
-    post_passed = sum(1 for r in post_qc_results if r["verdict"] == "PASS")
-    post_warned = sum(1 for r in post_qc_results if r["verdict"] == "WARN")
-    post_failed = sum(1 for r in post_qc_results if r["verdict"] == "FAIL")
-
-    for r in post_qc_results:
-        for check_name, check_info in r["checks"].items():
-            if check_info["status"] == "WARN":
-                warnings.append(f"Post-QC {r['filename']}: {check_name} WARN — {check_info['detail']}")
-
-    if post_failed > 0:
-        failed_tracks = [r for r in post_qc_results if r["verdict"] == "FAIL"]
-        fail_details = []
-        for r in failed_tracks:
-            for check_name, check_info in r["checks"].items():
-                if check_info["status"] == "FAIL":
-                    fail_details.append({
-                        "filename": r["filename"],
-                        "check": check_name,
-                        "status": "FAIL",
-                        "detail": check_info["detail"],
-                    })
-
-        stages["post_qc"] = {
-            "status": "fail",
-            "passed": post_passed,
-            "warned": post_warned,
-            "failed": post_failed,
-            "verdict": "FAILURES FOUND",
-        }
-        return _safe_json({
-            "album_slug": album_slug,
-            "stage_reached": "post_qc",
-            "stages": stages,
-            "settings": settings,
-            "warnings": warnings,
-            "failed_stage": "post_qc",
-            "failure_detail": {
-                "tracks_failed": [r["filename"] for r in failed_tracks],
-                "details": fail_details,
-            },
-        })
-
-    stages["post_qc"] = {
-        "status": "pass",
-        "passed": post_passed,
-        "warned": post_warned,
-        "failed": 0,
-        "verdict": "ALL PASS" if post_warned == 0 else "WARNINGS",
-    }
+    if result := await _album_stages._stage_post_qc(ctx):
+        return result
 
     # --- Stage 6.5: Archival (opt-in) ---
-    # When mastering.archival_enabled is true, write a 32-bit float copy
-    # of each mastered track to archival/. This is a bit-depth-expanded
-    # copy of the delivery master (not a separate render), intended for
-    # re-mastering without re-polishing stems.
-    #
-    # The archival dir is a MIRROR of mastered/: orphans (files whose
-    # basename is no longer in mastered/) are removed before the new
-    # float copies go down. Keeps re-mastered albums in sync when tracks
-    # are dropped or renamed (#290 phase 4 — PR #304 review item A5).
-    if targets.get("archival_enabled"):
-        import soundfile as _sf_archival
-        from tools.mastering.archival import prune_archival_orphans
-
-        archival_dir = audio_dir / "archival"
-        archival_dir.mkdir(exist_ok=True)
-        mastered_names = {p.name for p in mastered_files}
-        pruned = prune_archival_orphans(archival_dir, mastered_names)
-
-        archived = 0
-        archive_errors: list[str] = []
-        for mastered_path in mastered_files:
-            arch_path = archival_dir / mastered_path.name
-            try:
-                data, rate = _sf_archival.read(str(mastered_path), dtype="float32")
-                _sf_archival.write(str(arch_path), data, rate, subtype="FLOAT")
-                archived += 1
-            except Exception as exc:  # pragma: no cover - filesystem error path
-                archive_errors.append(f"{mastered_path.name}: {exc}")
-
-        stages["archival"] = {
-            "status": "pass" if not archive_errors else "warn",
-            "count": archived,
-            "pruned": pruned or None,
-            "output_dir": str(archival_dir),
-            "errors": archive_errors or None,
-        }
+    if result := await _album_stages._stage_archival(ctx):
+        return result
 
     # --- Stage 6.7: Album layout emitter (#290 step 7) ---
-    # Write LAYOUT.md to the audio dir (sibling of ALBUM_SIGNATURE.yaml)
-    # with one transition per adjacent track pair. MVP emits defaults —
-    # humans tune by hand-editing after generation.
-    layout_frontmatter = None
-    try:
-        _state_albums = (_shared.cache.get_state() or {}).get("albums", {})
-        _album_state = _state_albums.get(_normalize_slug(album_slug), {})
-        layout_frontmatter = _album_state.get("layout")
-    except Exception as _layout_state_exc:
-        logger.debug(
-            "Layout: could not read state.albums[%s].layout — %s. "
-            "Falling back to default transition.",
-            album_slug, _layout_state_exc,
-        )
-
-    default_transition = "gap"
-    if isinstance(layout_frontmatter, dict):
-        dt = layout_frontmatter.get("default_transition")
-        if dt in ("gap", "gapless"):
-            default_transition = dt
-
-    layout_stage: dict[str, Any] = {
-        "status":             "pass",
-        "path":               str(audio_dir / "LAYOUT.md"),
-        "default_transition": default_transition,
-        "transition_count":   0,
-    }
-    try:
-        track_filenames = [p.name for p in mastered_files]
-        transitions = _layout_compute_transitions(
-            track_filenames, default_transition=default_transition
-        )
-        layout_md = _layout_render_markdown(album_slug, transitions)
-        atomic_write_text(audio_dir / "LAYOUT.md", layout_md)
-        layout_stage["transition_count"] = len(transitions)
-    except (LayoutError, OSError) as exc:
-        layout_stage["status"] = "warn"
-        layout_stage["error"]  = str(exc)
-        warnings.append(f"Layout emitter: {exc}")
-
-    stages["layout"] = layout_stage
+    if result := await _album_stages._stage_layout(ctx):
+        return result
 
     # --- Stage 7: Update statuses ---
     state = _shared.cache.get_state_ref()
